requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
kafka-python==2.0.2
kafka-python-ng==2.2.3
newsapi-python==0.2.7
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.3
